# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Movie Booking Application.",
        "title": "Bioskop API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tickets/book": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Book seats for a specific schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ticketing"
                ],
                "summary": "Book tickets",
                "parameters": [
                    {
                        "description": "Booking Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/movie-app_internal_delivery_http_dto_request.BookTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/movie-app_pkg_utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/movie-app_internal_domain.Transaction"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Conflict / Double Booking",
                        "schema": {
                            "$ref": "#/definitions/movie-app_pkg_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/tickets/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all transaction history for current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ticketing"
                ],
                "summary": "Get booking history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/movie-app_pkg_utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/movie-app_internal_domain.Transaction"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tickets/schedules/{id}/seats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Check which seats are booked or free for a schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ticketing"
                ],
                "summary": "Get available seats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/movie-app_pkg_utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/movie-app_internal_delivery_http_dto_response.SeatAvailabilityResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/transactions/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancel a pending transaction manually",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Cancel transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/movie-app_pkg_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/transactions/{id}/pay": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Pay for a pending transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Pay transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment Method",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/movie-app_internal_delivery_http_dto_request.PayTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/movie-app_pkg_utils.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Already paid / Validation error",
                        "schema": {
                            "$ref": "#/definitions/movie-app_pkg_utils.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "movie-app_internal_delivery_http_dto_request.BookTicketRequest": {
            "type": "object",
            "required": [
                "schedule_id",
                "seat_ids"
            ],
            "properties": {
                "promo_code": {
                    "type": "string"
                },
                "schedule_id": {
                    "type": "string"
                },
                "seat_ids": {
                    "description": "Array of Seat UUID",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "movie-app_internal_delivery_http_dto_request.PayTransactionRequest": {
            "type": "object",
            "required": [
                "payment_method"
            ],
            "properties": {
                "payment_method": {
                    "type": "string",
                    "enum": [
                        "credit_card",
                        "e_wallet",
                        "qris"
                    ]
                }
            }
        },
...
        "movie-app_internal_delivery_http_dto_response.SeatAvailabilityResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_booked": {
                    "description": "True jika sudah ada yang punya",
                    "type": "boolean"
                },
                "row_code": {
                    "type": "string"
                },
                "seat_number": {
                    "type": "integer"
                }
            }
        },
...
        "movie-app_internal_domain.Promo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "discount_type": {
                    "description": "'percentage' or 'fixed'",
                    "type": "string"
                },
                "discount_value": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "valid_until": {
                    "type": "string"
                }
            }
        },
        "movie-app_internal_domain.Seat": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "row_code": {
                    "description": "A, B, C",
                    "type": "string"
                },
                "seat_number": {
                    "description": "1, 2, 3",
                    "type": "integer"
                },
                "studio_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "movie-app_internal_domain.Ticket": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "schedule_id": {
                    "type": "string"
                },
                "seat": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/movie-app_internal_domain.Seat"
                        }
                    ]
                },
                "seat_id": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "movie-app_internal_domain.Transaction": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "discount_amount": {
                    "type": "number"
                },
                "expires_at": {
                    "type": "string"
                },
                "final_amount": {
                    "description": "Total setelah diskon",
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "promo": {
                    "$ref": "#/definitions/movie-app_internal_domain.Promo"
                },
                "promo_id": {
                    "description": "--- Tambahan Field Promo ---",
                    "type": "string"
                },
                "reminder_sent": {
                    "type": "boolean"
                },
                "status": {
                    "$ref": "#/definitions/movie-app_internal_enums.TransactionStatus"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/movie-app_internal_domain.Ticket"
                    }
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/movie-app_internal_domain.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "movie-app_internal_domain.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "description": "Menggunakan Enum",
                    "allOf": [
                        {
                            "$ref": "#/definitions/movie-app_internal_enums.Role"
                        }
                    ]
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "movie-app_internal_enums.Role": {
            "type": "string",
            "enum": [
                "admin",
                "user"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleUser"
            ]
        },
        "movie-app_internal_enums.TransactionStatus": {
            "type": "string",
            "enum": [
                "pending",
                "paid",
                "cancelled",
                "failed"
            ],
            "x-enum-varnames": [
                "TransactionPending",
                "TransactionPaid",
                "TransactionCancel",
                "TransactionFailed"
            ]
        },
        "movie-app_pkg_utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}